Use type uint32_t for database key and data lengths and for related variables.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1496 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ datastore.h
@@ -64,8 +64,9 @@ long db_get_msgcount(void*);
 /** set the database message count */
 void db_set_msgcount(void*, long);
 
-typedef int (*db_foreach_t)(char *key, long key_size,
-	char *value, long key_value, void *userdata);
+typedef int (*db_foreach_t)(char *key,   uint32_t key_size,
+			    char *value, uint32_t key_value, 
+			    void *userdata);
 /** Iterate over all elements in data base and call \p hook for each item.
  * \p userdata is passed through to the hook function unaltered.
  */
